Support more lemmy v4 endpoints

+++ api/src/main/kotlin/com/idunnololz/summit/api/dto/lemmy/v4/models/ListMediaI.kt
@@ -29,7 +29,7 @@ import com.google.gson.annotations.SerializedName
 data class ListMediaI (
 
     @SerializedName("limit")
-    val limit: kotlin.Double? = null,
+    val limit: kotlin.Int? = null,
 
     /* To get the next or previous page, pass this string unchanged as `page_cursor` in a new request to the same endpoint.  Do not attempt to parse or modify the cursor string. The format is internal and may change in minor Lemmy versions. */
     @SerializedName("page_cursor")
